+++ winpr/libwinpr/utils/wlog/wlog.c
@@ -104,7 +104,7 @@ static BOOL log_recursion(LPCSTR file, LPCSTR fkt, int line)
 		return FALSE;
 
 	for (i=0; i<used; i++)
-		if (fprintf(stderr, "%s: %zd: %s\n", fkt, i, msg[i]) < 0)
+		if (fprintf(stderr, "%s: %lu: %s\n", fkt, (unsigned long)i, msg[i]) < 0)
 			return FALSE;
 
 #endif
